CMake based build system.
For now, this can be used alternatively to the current autoconf/make based
build system.  It offers more features already though, like a shared
library for our parser components which is used by the test and
main executable.  So the autoconf/make based system might as well get
removed, we will see.